coming together

+++ transferfuncproperty.h
@@ -14,7 +14,7 @@ public slots:
 	void colorChangedAt(int index, QColor newColor);
 private:
 	void updateTFTexture();
-	void getBlendedColors(float coeff, QColor c1, QColor c2, float& final_red, float& final_green, float& final_blue, float& final_alpha);
+	void getBlendedColors(float coeff, QColor c1, QColor c2, int& final_red, int& final_green, int& final_blue, int& final_alpha);
 	void init_keys();
 	QVector<std::shared_ptr<TransferFuncKey>> keys;
 };
